refactor: extract webview navigation helper and fix lint warnings
- Remove unnecessary imports in execute_handler.dart
- Extract back navigation JavaScript to WebViewNavigationHelper class
- Reduce form_gear_webview.dart from 20KB/507LOC to 12KB/365LOC
- Both files now under 18KB sanity check limit
